Make sure links to media are correct
