+++ lib/json_serialize.rb
@@ -50,13 +50,11 @@ end
 
 # validator to go along with JsonSerialize
 class MaxJsonLengthValidator < ActiveModel::EachValidator
-
   def validate_each(record, attribute, value)
     if record[attribute].to_s.length > options[:with]
       record.errors[attribute] << (options[:message] || "is too long")
     end
   end
-
 end
 
 ActiveRecord::Base.extend JsonSerialize::ClassMethods
